+++ doc/Sphinx/Overview/releases.rst
@@ -36,6 +36,7 @@ Changes made in the repository (not released)
 * Laser Envelope: multi-level tunnel ionization creates multiple electrons, improving the sampling
 * Happi:
 
+  * operations in ``Fields``, ``Probe``, and ``ParticleBinning`` may now accept physical constants or units
   * ``Probe`` has a new method ``changeField``
   * ``Open`` has a new argument ``pint`` to disable the *Pint* package